Prevent undefined shift intrinsics from being used when `R128_STDC_ONLY` is defined
`R128_STDC_ONLY` causes intrinsic headers to not be included. It can be defined manually (via build options) even on systems where those intrinsics are available. This patch takes care of a case where only the availability of the intrinsic is checked, and not whether it's force-disabled by `R128_STDC_ONLY`.